Implements enableConstraintsChecks
